+++ ruma-events/src/policy.rs
@@ -1,0 +1,3 @@
+//! Modules for events in the *m.policy* namespace.
+
+pub mod rule;